Migrate code to .NET 6
