+++ dotnet-desktop-guide/net/winforms/get-started/snippets/create-app-visual-studio/csharp/Form1.cs
@@ -1,0 +1,16 @@
+namespace Names
+{
+    public partial class Form1 : Form
+    {
+        public Form1()
+        {
+            InitializeComponent();
+        }
+
+        private void btnAdd_Click(object sender, EventArgs e)
+        {
+            if (!string.IsNullOrWhiteSpace(txtName.Text) && !lstNames.Items.Contains(txtName.Text))
+                lstNames.Items.Add(txtName.Text);
+        }
+    }
+}
