+++ dotnet-desktop-guide/net/winforms/get-started/snippets/create-app-visual-studio/csharp/Form1.resx
