+++ dotnet-desktop-guide/net/winforms/get-started/snippets/create-app-visual-studio/csharp/Names.csproj
@@ -2,13 +2,10 @@
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net5.0-windows</TargetFramework>
+    <TargetFramework>net6.0-windows</TargetFramework>
+    <Nullable>enable</Nullable>
     <UseWindowsForms>true</UseWindowsForms>
-    <StartupObject>Sub Main</StartupObject>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Update="Main.vb" />
-  </ItemGroup>
 
 </Project>
