--- dotnet-desktop-guide/net/winforms/get-started/snippets/create-app-visual-studio/csharp/project.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net5.0-windows</TargetFramework>
-    <UseWindowsForms>true</UseWindowsForms>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Update="Main.cs" />
-  </ItemGroup>
-
-</Project>
