+++ dotnet-desktop-guide/net/winforms/get-started/snippets/create-app-visual-studio/vb/Form1.resx
