+++ dotnet-desktop-guide/net/winforms/get-started/snippets/create-app-visual-studio/vb/Form1.vb
@@ -1,16 +1,4 @@
-﻿Imports System.Windows.Forms
-Imports System.Drawing
-
-Partial Public Class Form1
-
-    Public Sub New()
-        InitializeComponent()
-    End Sub
-
-    Private Sub Form1_Load(sender As Object, e As EventArgs) Handles MyBase.Load
-
-    End Sub
-
+﻿Public Class Form1
     Private Sub btnAdd_Click(sender As Object, e As EventArgs) Handles btnAdd.Click
         If Not String.IsNullOrWhiteSpace(txtName.Text) And Not lstNames.Items.Contains(txtName.Text) Then
             lstNames.Items.Add(txtName.Text)
