--- dotnet-desktop-guide/net/winforms/get-started/snippets/create-app-visual-studio/vb/Program.vb
@@ -1,15 +1,0 @@
-﻿Imports System
-Imports System.Collections.Generic
-Imports System.Linq
-Imports System.Threading.Tasks
-Imports System.Windows.Forms
-
-Module Program
-    Sub Main(args As String())
-        Application.SetHighDpiMode(HighDpiMode.SystemAware)
-        Application.EnableVisualStyles()
-        Application.SetCompatibleTextRenderingDefault(False)
-        Application.Run(New Form1())
-    End Sub
-
-End Module
